+++ website/models/talk.py
@@ -2,17 +2,26 @@ from django.db import models
 from django.conf import settings
 from django.db.models.signals import pre_delete, post_save, m2m_changed, post_delete
 from django.dispatch import receiver
+from django.utils.text import get_valid_filename
 
 from sortedm2m.fields import SortedManyToManyField
 
 import os
+import os.path
+import logging
 
 from .project_umbrella import Project_umbrella
 from .keyword import Keyword
 from .person import Person
 from .video import Video
 
+# This retrieves a Python logging instance (or creates it)
+_logger = logging.getLogger(__name__)
+
 class Talk(models.Model):
+    UPLOAD_DIR = 'talks/' # relative path
+    THUMBNAIL_DIR = os.path.join(UPLOAD_DIR, 'images/') # relative path
+
     title = models.CharField(max_length=255)
 
     # A talk can be about more than one project
@@ -37,8 +46,8 @@ class Talk(models.Model):
 
     # The PDF and raw files (e.g., keynote, pptx) are required
     # TODO: remove null=True from these two fields
-    pdf_file = models.FileField(upload_to='talks/', null=True, default=None, max_length=255)
-    raw_file = models.FileField(upload_to='talks/', blank=True, null=True, default=None, max_length=255)
+    pdf_file = models.FileField(upload_to=UPLOAD_DIR, null=True, default=None, max_length=255)
+    raw_file = models.FileField(upload_to=UPLOAD_DIR, blank=True, null=True, default=None, max_length=255)
 
     INVITED_TALK = "Invited Talk"
     CONFERENCE_TALK = "Conference Talk"
@@ -61,7 +70,7 @@ class Talk(models.Model):
     # The thumbnail should have null=True because it is added automatically later by a post_save signal
     # TODO: decide if we should have this be editable=True and if user doesn't add one him/herself, then
     # auto-generate thumbnail
-    thumbnail = models.ImageField(upload_to='talks/images/', editable=False, null=True, max_length=255)
+    thumbnail = models.ImageField(upload_to=THUMBNAIL_DIR, editable=False, null=True, max_length=255)
 
     # raw_file = models.FileField(upload_to='talks/')
     # print("In talk model!")
@@ -102,36 +111,26 @@ def update_file_name_talks(sender, instance, action, reverse, **kwargs):
         title = ''.join(x for x in instance.title.title() if not x.isspace())
         title = ''.join(e for e in title if e.isalnum())
 
-        #change the pdf_file path to point to the renamed file
-        instance.pdf_file.name = os.path.join('talks', name + '_' + title + '_' + str(year) + '.pdf')
+        # Convert metadata into a filename
+        new_filename = name + '_' + title + '_' + instance.forum_name + '_' + str(year) + '.pdf'
+
+        # Use Django helper function to ensure a clean filename
+        new_filename = get_valid_filename(new_filename)
+
+        # Change the pdf_file path to point to the renamed file
+        instance.pdf_file.name = os.path.join(Talk.UPLOAD_DIR, new_filename)
+
+        # Add in the media directory
         new_path = os.path.join(settings.MEDIA_ROOT, instance.pdf_file.name)
-        os.rename(initial_path, new_path)
-        instance.save()
+        
+        # Actually rename the existing file (aka initial_path) but only if it exists (it should!)
+        if os.path.exists(initial_path):
+            os.rename(initial_path, new_path)
+            instance.save()
+        else:
+            _logger.error(f'The file {initial_path} does not exist and cannot be renamed to {new_path}')
 
-        # old_pdf_filename_with_path = instance.pdf_file.path
-        # new_pdf_filename = get_formatted_filename(instance.get_person(), instance.title, instance.date.year, '.pdf')
-        #
-        # old_raw_filename_with_path = instance.raw_file.path
-        # raw_ext = os.path.splitext(instance.raw_file.path)[1]
-        # new_raw_filename = get_formatted_filename(instance.get_person(), instance.title, instance.date.year, raw_ext)
-
-# def get_formatted_filename(person, title_str, year, ext):
-#     initial_path = instance.pdf_file.path
-#     person = instance.get_person()
-#     name = person.last_name
-#     year = instance.date.year
-#     title = instance.title.title()
-#     title = ''.join(x for x in title if not x.isspace())
-#     title = ''.join(e for e in title if e.isalnum())
-#
-#     # change the pdf_file path to point to the renamed file
-#     instance.pdf_file.name = os.path.join(dir_name, name + '_' + title + '_' + str(year) + '.pdf')
-#     new_path = os.path.join(settings.MEDIA_ROOT, instance.pdf_file.name)
-#     os.rename(initial_path, new_path)
-#     instance.save()
-
-
-
+        
 m2m_changed.connect(update_file_name_talks, sender=Talk.speakers.through)
 
 @receiver(post_delete, sender=Talk)
